// src/types/lpc-assets.ts
export const LPC_SHEET_DEFINITION_GROUPS = [
  "arms",
  "body",
  "feet",
  "hair",
  "head",
  "headwear",
  "legs",
  "tools",
  "torso",
  "weapons"
] as const;

export type LpcSheetDefinitionGroup = (typeof LPC_SHEET_DEFINITION_GROUPS)[number];